requirements.txt
openai
python-dotenv
black
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
oauth2client
httpx